move get Post language
